refined text to properly scale with screen width. further refinement needed.

+++ src/main/java/garbageboys/garbageman_mk_2/Text/TextManager.java
@@ -15,8 +15,8 @@ public interface TextManager {
 						float size,
 						float x,
 						float y,
-						int max_height, 
-						int width 
+						float max_height, 
+						float width 
 						);
 	
 	public void closeText(TextObject text_object);
